Controller layer, exception handling, swagger

+++ driver_service/src/main/java/com/internship/driver_service/repo/RateRepo.java
@@ -9,6 +9,5 @@ public interface RateRepo extends JpaRepository<Rate, Long> {
     @Query("select FLOOR(AVG(r.value)) " +
             "from Rate r " +
             "where r.driver.profileId=:profileId ")
-    Byte findDriverRatingByProfileId(@Param("profileId") Long profileId);
-
+    Integer findDriverRatingByProfileId(@Param("profileId") Long profileId);
 }
